# rl/plugins/VIM/kubernetes/mtp_plugin_kubernetes/swagger/swagger.yaml
...
swagger: "2.0"
...
info:
  description: "MTP Manager API"
  version: "1.0"
  title: "MTP Manager API"
  termsOfService: "/tos"
paths:
  /compute_resources:
    get:
      tags:
      - "virtualisedComputeResources"
      operationId: "query_resources"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "computeId"
        in: "query"
        description: "Identifier(s) of the abstract compute resource(s) to be terminated."
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "Element containing information about the virtual compute resource(s)\
            \ matching the filter. The cardinality can be 0 if no matching compute\
            \ resources exist."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VirtualCompute"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
    post:
      tags:
      - "virtualisedComputeResources"
      operationId: "allocate_compute"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/AllocateComputeRequest"
      responses:
        201:
          description: "Element containing information of the newly instantiated virtualised\
            \ compute resource."
          schema:
            $ref: "#/definitions/VirtualCompute"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        409:
          description: "software image already added"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
    delete:
      tags:
      - "virtualisedComputeResources"
      operationId: "terminate_abstract_resources"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "computeId"
        in: "query"
        description: "Identifier(s) of the abstract compute resource(s) to be terminated."
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "Identifier(s) of the virtualised compute resource(s) successfully\
            \ terminated."
          schema:
            type: "array"
            items:
              type: "string"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
  /compute_resources/capacities:
    get:
      tags:
      - "virtualisedComputeResources"
      operationId: "query_compute_capacity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "ComputeResourceTypeId"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "The capacity during the requested time period. The scope is\
            \ according to parameter zoneId of the request during the time interval."
          schema:
            $ref: "#/definitions/CapacityInformation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
  /compute_resources/flavours:
    get:
      tags:
      - "virtualisedComputeResources"
      operationId: "query_flavors"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Element containing information about the virtual compute resource(s)\
            \ matching the filter. The cardinality can be 0 if no matching compute\
            \ resources exist."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VirtualComputeFlavour"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
    post:
      tags:
      - "virtualisedComputeResources"
      operationId: "create_flavour"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Flavour"
        required: true
        schema:
          $ref: "#/definitions/VirtualComputeFlavour"
      responses:
        201:
          description: "Identifier of the created Compute Flavour."
          schema:
            type: "string"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        409:
          description: "flavour already added"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
  /compute_resources/flavours/{id}:
    delete:
      tags:
      - "virtualisedComputeResources"
      operationId: "delete_flavours"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Compute Flavour to be deleted."
        required: true
        type: "string"
      responses:
        200:
          description: "No output parameters"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
  /compute_resources/information:
    get:
      tags:
      - "virtualisedComputeResources"
      operationId: "query_compute_information"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "query"
        description: "Filter defining the information of consumable virtualised resources\
          \ on which the query applies."
        required: true
        type: "string"
      responses:
        200:
          description: "Virtualised compute resource information in the VIM that satisfies\
            \ the query condition."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VirtualComputeResourceInformation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
  /compute_resources/nfvi_pop_compute_information:
    get:
      tags:
      - "virtualisedComputeResources"
      operationId: "query_nfvi_po_p_compute_information"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "The filtered information that has been retrieved. The cardinality\
            \ can be 0 if no matching information exist."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/NfviPop"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
  /compute_resources/resource_zones:
    get:
      tags:
      - "virtualisedComputeResources"
      operationId: "query_compute_resource_zone"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "The filtered information that has been retrieved about the\
            \ Resource Zone. The cardinality can be 0 if no matching information exist."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ResourceZone"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_compute_resources_controller"
...
  /information:
    get:
      tags:
      - "virtualisedNetworkResourcesInformation"
      operationId: "query_network_information"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "informationQueryFilter"
        description: "Filter defining the information of consumable virtualised resources\
          \ on which the query applies."
        required: true
        schema:
          $ref: "#/definitions/Filter"
      responses:
        200:
          description: "Virtualised network resource information in the VIM that satisfies\
            \ the query condition."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VirtualNetworkResourceInformation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_network_resources_information_controller"
...
definitions:
...
  AllocateComputeRequest:
    type: "object"
    required:
    - "affinityOrAntiAffinityConstraints"
    - "computeFlavourId"
    - "computeName"
    - "interfaceData"
    - "locationConstraints"
    - "metadata"
    - "reservationId"
    - "resourceGroupId"
    - "userData"
    - "vcImageId"
    properties:
      affinityOrAntiAffinityConstraints:
        type: "array"
        description: "A list of elements with affinity or anti affinity information\
          \ of the virtualised compute resource to allocate. There should be a relation\
          \ between the constraints defined in the different elements of the list."
        items:
          $ref: "#/definitions/AllocateComputeRequest_affinityOrAntiAffinityConstraints"
      computeFlavourId:
        type: "string"
        description: "Identifier of the Compute Flavour that provides information\
          \ about the particular memory, CPU and disk resources for virtualised compute\
          \ resource to allocate. The Compute Flavour is created with Create Compute\
          \ Flavour operation"
      computeName:
        type: "string"
        description: "Name provided by the consumer for the virtualised compute resource\
          \ to allocate. It can be used for identifying resources from consumer side."
      interfaceData:
        type: "array"
        description: "The data of network interfaces which are specific to a Virtual\
          \ Compute Resource instance."
        items:
          $ref: "#/definitions/AllocateComputeRequest_interfaceData"
      locationConstraints:
        type: "string"
        description: "If present, it defines location constraints for the resource(s)\
          \ is (are) requested to be allocated, e.g. in what particular Resource Zone."
      metadata:
        type: "array"
        description: "The binary software image file."
        items:
          $ref: "#/definitions/metaData_inner"
      reservationId:
        type: "string"
        description: "Identifier of the resource reservation applicable to this virtualised\
          \ resource management operation. Cardinality can be 0 if no resource reservation\
          \ was used."
      resourceGroupId:
        type: "string"
        description: "Unique identifier of the \"infrastructure resource group\",\
          \ logical grouping of virtual resources assigned to a tenant within an Infrastructure\
          \ Domain."
      userData:
        $ref: "#/definitions/AllocateComputeRequest_userData"
      vcImageId:
        type: "string"
        description: "Identifier of the virtualisation container software image (e.g.\
          \ a virtual machine image). Cardinality can be 0 if an \"empty\" virtualisation\
          \ container is allocated. "
    example:
      metadata:
      - value: "value"
        key: "key"
      - value: "value"
        key: "key"
      userData:
        method: "method"
        content: "content"
      affinityOrAntiAffinityConstraints:
      - affinityAntiAffinityResourceGroup: "affinityAntiAffinityResourceGroup"
        scope: "scope"
        type: "type"
        affinityAntiAffinityResourceList:
          resource:
          - "resource"
          - "resource"
      - affinityAntiAffinityResourceGroup: "affinityAntiAffinityResourceGroup"
        scope: "scope"
        type: "type"
        affinityAntiAffinityResourceList:
          resource:
          - "resource"
          - "resource"
      reservationId: "reservationId"
      vcImageId: "vcImageId"
      locationConstraints: "locationConstraints"
      interfaceData:
      - macAddress: "macAddress"
        ipAddress: "ipAddress"
      - macAddress: "macAddress"
        ipAddress: "ipAddress"
      resourceGroupId: "resourceGroupId"
      computeFlavourId: "computeFlavourId"
      computeName: "computeName"
...
  CapacityInformation:
    type: "object"
    required:
    - "allocatedCapacity"
    - "availableCapacity"
    - "reservedCapacity"
    - "totalCapacity"
    properties:
      allocatedCapacity:
        type: "string"
        description: "The allocated capacity is usually specified as the current allocated\
          \ capacity."
      availableCapacity:
        type: "string"
        description: "The free capacity available for allocation and reservation.\
          \ It can be specified in terms of current capacity; or minimum and maximum\
          \ capacity; average capacity; or other statistical measurement in the specified\
          \ time interval. The set of measurements is to be defined during Stage 3."
      reservedCapacity:
        type: "string"
        description: "The reserved capacity. It can be specified in terms of current\
          \ capacity; or minimum and maximum capacity; average capacity; or other\
          \ statistical measurement in the specified time interval. The set of measurements\
          \ is to be defined during Stage 3."
      totalCapacity:
        type: "string"
        description: "The total capacity is usually specified as a fixed capacity\
          \ without variations in time. The set of measurements is left to Stage 3."
    example:
      availableCapacity: "availableCapacity"
      reservedCapacity: "reservedCapacity"
      allocatedCapacity: "allocatedCapacity"
      totalCapacity: "totalCapacity"
...
  NfviPop:
    type: "object"
    required:
    - "geographicalLocationInfo"
    - "networkConnectivityEndpoint"
    - "nfviPopId"
    - "vimId"
    properties:
      geographicalLocationInfo:
        type: "string"
        description: "It provides information about the geographic location (e.g.\
          \ geographic coordinates or address of the building, etc.) of the NFVI resources\
          \ that the VIM manages."
      networkConnectivityEndpoint:
        type: "string"
        description: "Information about network connectivity endpoints to the NFVI-PoP\
          \ that the VIM manages which helps build topology information relative to\
          \ NFVI-PoP connectivity to other NFVI-PoP or N-PoP."
      nfviPopId:
        type: "string"
        description: "Identification of the NFVI-PoP."
      vimId:
        type: "string"
        description: "Identification of the VIM."
    example:
      networkConnectivityEndpoint: "networkConnectivityEndpoint"
      vimId: "vimId"
      nfviPopId: "nfviPopId"
      geographicalLocationInfo: "geographicalLocationInfo"
...
  VirtualCompute:
    type: "object"
    required:
    - "accelerationCapability"
    - "computeId"
    - "computeName"
    - "flavourId"
    - "hostId"
    - "operationalState"
    - "vcImageId"
    - "virtualCpu"
    - "virtualDisks"
    - "virtualMemory"
    - "virtualNetworkInterface"
    - "zoneId"
    properties:
      accelerationCapability:
        type: "array"
        description: "Selected acceleration capabilities (e.g. crypto, GPU) from the\
          \ set of capabilities offered by the compute node acceleration resources.\
          \ The cardinality can be 0, if no particular acceleration capability is\
          \ provided."
        items:
          type: "string"
      computeId:
        type: "string"
        description: "Identifier of the virtualised compute resource."
      computeName:
        type: "string"
        description: "Name of the virtualised compute resource."
      flavourId:
        type: "string"
        description: "Identifier of the given compute flavour used to instantiate\
          \ this virtual compute."
      hostId:
        type: "string"
        description: "Identifier of the host the virtualised compute resource is allocated\
          \ on."
      operationalState:
        type: "string"
        description: "Operational state of the compute resource."
      vcImageId:
        type: "string"
        description: "Identifier of the virtualisation container software image (e.g.\
          \ virtual machine image). Cardinality can be 0 if an \"empty\" virtualisation\
          \ container is allocated."
      virtualCpu:
        $ref: "#/definitions/VirtualCompute_virtualCpu"
      virtualDisks:
        type: "string"
        description: "Element with information of the virtualised storage resources\
          \ (volumes, ephemeral that are attached to the compute resource.)"
      virtualMemory:
        $ref: "#/definitions/VirtualCompute_virtualMemory"
      virtualNetworkInterface:
        type: "array"
        description: "Element with information of the instantiated virtual network\
          \ interfaces of the compute resource."
        items:
          $ref: "#/definitions/ReservedVirtualCompute_virtualisationContainerReserved_virtualNetworkInterface"
      zoneId:
        type: "string"
        description: "If present, it identifies the Resource Zone where the virtual\
          \ compute resources have been allocated."
    example:
      flavourId: "flavourId"
      computeId: "computeId"
      vcImageId: "vcImageId"
      accelerationCapability:
      - "accelerationCapability"
      - "accelerationCapability"
      hostId: "hostId"
      virtualCpu:
        numVirtualCpu: 6
        virtualCpuOversubscriptionPolicy: "virtualCpuOversubscriptionPolicy"
        cpuArchitecture: "cpuArchitecture"
        virtualCpuPinning:
          cpuMap: "cpuMap"
          cpuPinningRules: "cpuPinningRules"
          cpuPinningPolicy: "cpuPinningPolicy"
        cpuClock: 0.80082819046101150206595775671303272247314453125
      zoneId: "zoneId"
      operationalState: "operationalState"
      virtualDisks: "virtualDisks"
      computeName: "computeName"
      virtualMemory:
        numaEnabled: true
        virtualMemSize: 1.46581298050294517310021547018550336360931396484375
        virtualMemOversubscriptionPolicy: "virtualMemOversubscriptionPolicy"
      virtualNetworkInterface:
      - typeConfiguration: "typeConfiguration"
        typeVirtualNic: "typeVirtualNic"
        macAddress: "macAddress"
        metadata: "metadata"
        resourceId: "resourceId"
        networkPortId: "networkPortId"
        accelerationCapability: "accelerationCapability"
        bandwidth: "bandwidth"
        ipAddress: "ipAddress"
        networkId: "networkId"
        operationalState: "operationalState"
        ownerId: "ownerId"
      - typeConfiguration: "typeConfiguration"
        typeVirtualNic: "typeVirtualNic"
        macAddress: "macAddress"
        metadata: "metadata"
        resourceId: "resourceId"
        networkPortId: "networkPortId"
        accelerationCapability: "accelerationCapability"
        bandwidth: "bandwidth"
        ipAddress: "ipAddress"
        networkId: "networkId"
        operationalState: "operationalState"
        ownerId: "ownerId"
  VirtualComputeFlavour:
    type: "object"
    required:
    - "accelerationCapability"
    - "flavourId"
    - "storageAttributes"
    - "virtualCpu"
    - "virtualMemory"
    - "virtualNetworkInterface"
    properties:
      accelerationCapability:
        type: "string"
        description: "Selected acceleration capabilities (e.g. crypto, GPU) from the\
          \ set of capabilities offered by the compute node acceleration resources.\
          \ The cardinality can be 0, if no particular acceleration capability is\
          \ requested."
      flavourId:
        type: "string"
        description: "Identifier given to the compute flavour."
      storageAttributes:
        type: "array"
        description: "Element containing information about the size of virtualised\
          \ storage resources (e.g. size of volume, in GB), the type of storage (e.g.\
          \ volume, object)."
        items:
          $ref: "#/definitions/CreateComputeResourceReservationRequest_containerFlavour_storageAttributes"
      virtualCpu:
        type: "array"
        description: "The virtual CPU(s) of the virtualised compute."
        items:
          $ref: "#/definitions/CreateComputeResourceReservationRequest_containerFlavour_virtualCpu"
      virtualMemory:
        $ref: "#/definitions/CreateComputeResourceReservationRequest_containerFlavour_virtualMemory"
      virtualNetworkInterface:
        $ref: "#/definitions/CreateComputeResourceReservationRequest_containerFlavour_virtualNetworkInterface"
    example:
      flavourId: "flavourId"
      accelerationCapability: "accelerationCapability"
      storageAttributes:
      - flavourId: "flavourId"
        typeOfStorage: "typeOfStorage"
        hostId: "hostId"
        zoneId: "zoneId"
        operationalState: "operationalState"
        ownerId: "ownerId"
        sizeOfStorage: 0.80082819046101150206595775671303272247314453125
        storageId: "storageId"
        storageName: "storageName"
      - flavourId: "flavourId"
        typeOfStorage: "typeOfStorage"
        hostId: "hostId"
        zoneId: "zoneId"
        operationalState: "operationalState"
        ownerId: "ownerId"
        sizeOfStorage: 0.80082819046101150206595775671303272247314453125
        storageId: "storageId"
        storageName: "storageName"
      virtualCpu:
      - numVirtualCpu: 1
        virtualCpuOversubscriptionPolicy: "virtualCpuOversubscriptionPolicy"
        cpuArchitecture: "cpuArchitecture"
        virtualCpuPinning:
          cpuMap: "cpuMap"
          cpuPinningRules: "cpuPinningRules"
          cpuPinningPolicy: "cpuPinningPolicy"
        cpuClock: 6.02745618307040320615897144307382404804229736328125
      - numVirtualCpu: 1
        virtualCpuOversubscriptionPolicy: "virtualCpuOversubscriptionPolicy"
        cpuArchitecture: "cpuArchitecture"
        virtualCpuPinning:
          cpuMap: "cpuMap"
          cpuPinningRules: "cpuPinningRules"
          cpuPinningPolicy: "cpuPinningPolicy"
        cpuClock: 6.02745618307040320615897144307382404804229736328125
      virtualMemory:
        numaEnabled: true
        virtualMemSize: 5.962133916683182377482808078639209270477294921875
        virtualMemOversubscriptionPolicy: "virtualMemOversubscriptionPolicy"
      virtualNetworkInterface:
        typeConfiguration: "typeConfiguration"
        typeVirtualNic: "typeVirtualNic"
        macAddress: "macAddress"
        metadata: "metadata"
        resourceId: "resourceId"
        networkPortId: "networkPortId"
        accelerationCapability: "accelerationCapability"
        bandwidth: "bandwidth"
        ipAddress: "ipAddress"
        networkId: "networkId"
        operationalState: "operationalState"
        ownerId: "ownerId"
...
  VirtualComputeResourceInformation:
    type: "object"
    required:
    - "accelerationCapability"
    - "computeResourceTypeId"
    - "virtualCPU"
    - "virtualMemory"
    properties:
      accelerationCapability:
        type: "string"
        description: "Acceleration capabilities (e.g. crypto, GPU) for the consumable\
          \ virtualised compute resources from the set of capabilities offered by\
          \ the compute node acceleration resources. The cardinality can be 0, if\
          \ no particular acceleration capability is provided."
      computeResourceTypeId:
        type: "string"
        description: "Identifier of the consumable virtualised compute resource type."
      virtualCPU:
        $ref: "#/definitions/VirtualComputeResourceInformation_virtualCPU"
      virtualMemory:
        $ref: "#/definitions/VirtualComputeResourceInformation_virtualMemory"
    example:
      computeResourceTypeId: "computeResourceTypeId"
      accelerationCapability: "accelerationCapability"
      virtualCPU:
        numVirtualCpu: 6.02745618307040320615897144307382404804229736328125
        virtualCpuOversubscriptionPolicy: "virtualCpuOversubscriptionPolicy"
        cpuArchitecture: "cpuArchitecture"
        cpuClock: 0.80082819046101150206595775671303272247314453125
        virtualCpuPinningSupported: true
      virtualMemory:
        numaSupported: true
        virtualMemSize: 1.46581298050294517310021547018550336360931396484375
        virtualMemOversubscriptionPolicy: "virtualMemOversubscriptionPolicy"
...